+++ docs/source/release_notes.rst
@@ -92,6 +92,8 @@ Availability: `AutomationHub`_, `Galaxy`_, `GitHub`_
 * Dev: Added a final run of the "safety" and "bandit" tools to the "publish"
   workflow, to ensure that a release cannot happen with any findings.
 
+* Docs: Improved the short descriptions of the modules. (issue #998)
+
 **Cleanup:**
 
 * Modernized the code to match the minimum Python version 3.8 (use of f-strings,